feat: Implement "Add book" using search query via API
Refactored out the search as SearchComponent which is able to search for
books independently, but uses the parent functionality to move books
into shelves.
Since the search API returns unshelved books with {shelf: null}, I've
introduced a sentinel array in the parent state (a "none" shelf) which
is always empty to make life easier.
There still is an issue that even if books are persisted into shelves,
which you can see by pressing F5, they still come back as {shelf: null}
from the search - this needs some more investigation or a hack to fill
the shelf info from parent state right after the items come back from
the search or maybe right before the shelf select control is opened.